feat: rename photopreview to confirm register and add it to the flow
