+++ app/src/pages/AuthPages/ConfirmRegister/styles.js
@@ -1,6 +1,6 @@
 import { StyleSheet } from 'react-native';
-import fonts from '../../../../../assets/styles/fontVariable';
-import colors from '../../../../../assets/styles/colorVariables';
+import fonts from '../../../../assets/styles/fontVariable';
+import colors from '../../../../assets/styles/colorVariables';
 
 const styles = StyleSheet.create({
     container: {
