+++ app/src/pages/AuthPages/RiskGroup/index.js
@@ -1,18 +1,12 @@
 import React, { useState } from 'react';
-import { View, Text, TouchableOpacity, ActivityIndicator } from 'react-native';
+import { View, Text, TouchableOpacity } from 'react-native';
 import Button from '../../../components/UI/button';
 import styles from './styles';
-import SessionService from '../../../services/Session';
 import { Icon } from 'react-native-elements';
-import colors from '../../../../assets/styles/colorVariables';
 import riskGroups from '../../../utils/riskGroupsObject';
-import { alertSuccess } from '../../../utils/Alert';
-import useService from '../../../services/useService';
 
 export default function RiskGroup({ route, navigation }) {
     const { userDataFromPersonalPage } = route.params;
-    const [loadingUserRegistration, setLoadingUserRegistration] =
-        useState(false);
     const [disease, setDisease] = useState({
         dc: false,
         hiv: false,
@@ -40,17 +34,7 @@ export default function RiskGroup({ route, navigation }) {
             ...userDataFromPersonalPage,
             riskGroup: newDisease,
         };
-
-        setLoadingUserRegistration(true);
-        const completeRegistration = await useService(
-            SessionService,
-            'signUp',
-            [completeRegistragionData],
-        );
-        if (!completeRegistration.error) {
-            navigation.navigate('login');
-            alertSuccess('Seu cadastro foi realizado com sucesso');
-        }
+        navigation.navigate('confirmRegister', { completeRegistragionData });
     };
 
     const renderPageHeader = () => (
@@ -71,10 +55,6 @@ export default function RiskGroup({ route, navigation }) {
                 </Text>
             </View>
         </>
-    );
-
-    const renderLoadingIndicator = () => (
-        <ActivityIndicator size="large" color={colors.primary} />
     );
 
     const renderRiskGroupSelection = () => {
@@ -107,16 +87,12 @@ export default function RiskGroup({ route, navigation }) {
             {renderPageHeader()}
             {renderRiskGroupSelection()}
             <View style={styles.btnView}>
-                {loadingUserRegistration ? (
-                    renderLoadingIndicator()
-                ) : (
-                    <Button
-                        disabled={false}
-                        title="Concluir"
-                        large
-                        press={confirmSignUp}
-                    />
-                )}
+                <Button
+                    disabled={false}
+                    title="Continuar"
+                    large
+                    press={confirmSignUp}
+                />
             </View>
         </View>
     );
